fix(ci): preload numpy before pytest-cov traces
Coverage starting before the first numpy import reloads
_multiarray_umath and leaves ndarray.sum broken. Register a
pytest11 plugin so numpy is a process singleton first. Drop
pythonpath=['.'] so openmw is imported once.

+++ OpenMW/openmw/prefetch_sparsity.py
@@ -62,7 +62,7 @@ def build_hot_neuron_index(
 
     active = activations > 0
     token_count = activations.shape[0]
-    freq = active.sum(axis=0) / float(token_count)
+    freq = active.astype(np.float64).sum(axis=0) / float(token_count)
 
     index = HotNeuronIndex()
     for layer in range(freq.shape[0]):
